[jax2tf] Fix infinite recursion when converting some bfloat16 arrays
This fix takes advantage of a new change in TF to share the bfloat16
dtype definition with JAX

Fixes: #5106
Relates to: #5108
