+++ jax/experimental/jax2tf/jax2tf.py
@@ -77,30 +77,12 @@ def _is_tfval(v: TfVal) -> bool:
     return False
 
 def _safe_convert_to_tensor(val, dtype=None) -> TfVal:
-  """Converts val to a Tensor.
-
-  This method wraps TensorFlow's `convert_to_tensor
-  <https://www.tensorflow.org/api_docs/python/tf/convert_to_tensor>`_ operator with
-  special case handling for when `val` is an instance of `jnp.bfloat16` or has a
-  `jnp.bfloat16` dtype. Because this type is not supported in numpy and different
-  from `tf.bfloat16.as_numpy_dtype`, `tf.convert_to_tensor` runs into trouble when
-  trying to convert it. In such a case, we solve the problem by viewing val as a
-  `ndarray` with a `uint16` dtype, for which conversion is properly defined. Then, we
-  simply bitcast it back to `bfloat16`.
-  """
   dtype = dtype if dtype else (val.dtype if hasattr(val, "dtype") else None)
-  if (dtype == jnp.bfloat16 or isinstance(val, jnp.bfloat16)):
-    if not isinstance(val, jnp.ndarray):
-      val = np.array(val, jnp.bfloat16)
+  conversion_type = to_tf_dtype(dtype) if dtype else None
+  # We can convert directly, because all dtypes (even bfloat16) are the same
+  # in JAX and TF.
+  return tf.convert_to_tensor(val, dtype=conversion_type)
 
-    val = tf.bitcast(tf.convert_to_tensor(val.view(jnp.uint16),
-                                          dtype=to_tf_dtype(jnp.uint16)),
-                     type=to_tf_dtype(jnp.bfloat16))
-  else:
-    conversion_type = to_tf_dtype(dtype) if dtype else None
-    val = tf.convert_to_tensor(val, dtype=conversion_type)
-
-  return val
 
 # The implementation rules for primitives. The rule will be called with the
 # arguments (TfVal) and must return TfVal (or a sequence thereof,
